requirements.txt
ctransformers==0.2.5
sentence-transformers==2.2.2
pinecone-client
langchain==0.2.0
flask
pypdf
python-dotenv
langchain-pinecone
langchain-community
openai
-e .
